+++ Otimizacao.h
@@ -23,10 +23,9 @@ public:
 
 
 /*cenario 2*/
-    //int max(int x, int y); //retorna o max entre x e y
     MaxHeap<int, int> makemaxheap_recompensa(); //max heap com prioridade é a recompensa / espaço ocupado pela encomenda - espaço será a soma do voluma e do peso da encomenda
     MinHeap<int, int> makeminheap_custo(); //min heap com prioridade é o custo / espaço da carrinha - espaço será a soma do voluma e do peso
-    float cenario2(); //retorna o valor do lucro máximo
+    int cenario2(); //retorna o valor do lucro máximo
 
 /*cenario 3*/
     MinHeap<int, int> makeminheap();
